refactor: separated ENUM, no Storiges in Storiges except FilmStorage

+++ src/main/java/ru/yandex/practicum/filmorate/service/DirectorService.java
@@ -2,10 +2,12 @@ package ru.yandex.practicum.filmorate.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.exception.WrongIdException;
 import ru.yandex.practicum.filmorate.model.Director;
 import ru.yandex.practicum.filmorate.storage.DirectorStorage;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -13,7 +15,14 @@ public class DirectorService {
     private final DirectorStorage directorStorage;
 
     public Director getDirectorById(int id) {
-        return directorStorage.getDirectorById(id);
+        if (isIncorrectId(id)) {
+            throw new WrongIdException("Param must be more then 0");
+        }
+        Optional<Director> directorOpt = directorStorage.getDirectorById(id);
+        if (directorOpt.isEmpty()) {
+            throw new WrongIdException("No director with id = " + id + " in DB was found.");
+        }
+        return directorOpt.get();
     }
 
     public List<Director> getDirectors() {
@@ -25,10 +34,24 @@ public class DirectorService {
     }
 
     public Director updateDirector(Director director) {
+        if (!isLegalDirectorId(director.getId())) {
+            return director;
+        }
         return directorStorage.updateDirector(director);
     }
 
     public long deleteDirector(long id) {
+        if (isIncorrectId(id))  {
+            throw new WrongIdException("Param must be more then 0");
+        }
         return directorStorage.deleteDirector(id);
+    }
+
+    public boolean isLegalDirectorId(int directorId) {
+        return getDirectorById(directorId) != null;
+    }
+
+    private boolean isIncorrectId(long id) {
+        return id <= 0;
     }
 }
