+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -3,6 +3,8 @@ package ru.yandex.practicum.filmorate.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.exception.ValidationException;
+import ru.yandex.practicum.filmorate.exception.WrongIdException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.Feed;
 import ru.yandex.practicum.filmorate.model.User;
@@ -11,7 +13,10 @@ import ru.yandex.practicum.filmorate.storage.FeedStorage;
 import ru.yandex.practicum.filmorate.storage.FriendStorage;
 import ru.yandex.practicum.filmorate.storage.UserStorage;
 
+import java.time.LocalDate;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Slf4j
 @Service
@@ -23,11 +28,22 @@ public class UserService {
     private final FeedStorage feedStorage;
 
     public User create(User user) {
+        changeNameToLogin(user);
+        if (isNotValid(user)) {
+            throw new ValidationException("Can't create new user. Check your data.");
+        }
         userStorage.add(user);
         return user;
     }
 
     public User update(User user) {
+        changeNameToLogin(user);
+        if (isNotValid(user)) {
+            throw new ValidationException("Can't create new user. Check your data.");
+        }
+        if (!isLegalUserId(user.getId())) {
+            return user;
+        }
         return userStorage.update(user);
     }
 
@@ -36,38 +52,89 @@ public class UserService {
     }
 
     public void deleteUserById(long id) {
+        if (isIncorrectId(id))  {
+            throw new WrongIdException("Param must be more then 0");
+        }
         userStorage.delete(id);
     }
 
     public void addFriend(long userId, long friendId) {
-        friendStorage.addFriend(userId, friendId);
+        if (isLegalUserId(userId) && isLegalUserId(friendId)) {
+            friendStorage.addFriend(userId, friendId);
+            feedStorage.addFriendRequest(userId, friendId);
+        }
     }
 
     public void deleteFriend(long userId, long friendId) {
-        friendStorage.deleteFriend(userId, friendId);
+        if (isLegalUserId(userId) && isLegalUserId(friendId)) {
+            friendStorage.deleteFriend(userId, friendId);
+            feedStorage.deleteFriendRequest(userId, friendId);
+        }
     }
 
     public void updateFriendRequest(long userId, long friendId) {
-        friendStorage.acceptFriendRequest(userId, friendId);
+        if (isLegalUserId(userId) && isLegalUserId(friendId)) {
+            friendStorage.acceptFriendRequest(userId, friendId);
+            feedStorage.acceptFriendRequest(userId, friendId);
+        }
     }
 
     public List<User> findCommonFriends(long userId, long otherId) {
+        if (!isLegalUserId(userId) || !isLegalUserId(otherId)) {
+            return new ArrayList<>();
+        }
         return userStorage.getCommonFriendsByUserId(userId, otherId);
     }
 
     public List<Feed> getEventsList(long userId) {
+        if (!isLegalUserId(userId)) {
+            return new ArrayList<>();
+        }
         return feedStorage.getFeedList(userId);
     }
 
     public User getById(long userId) {
-        return userStorage.getById(userId);
+        if (isIncorrectId(userId)) {
+            throw new WrongIdException("Param must be more then 0");
+        }
+        Optional<User> userOpt = userStorage.getById(userId);
+        if (userOpt.isEmpty()) {
+            throw new WrongIdException("No user with id = " + userId + " in DB was found.");
+        }
+        return userOpt.get();
     }
 
     public List<User> getFriends(long userId) {
+        if (!isLegalUserId(userId)) {
+            return new ArrayList<>();
+        }
         return userStorage.getFriendsByUserId(userId);
     }
 
     public List<Film> getRecommendations(long userId) {
+        if (!isLegalUserId(userId)) {
+            return new ArrayList<>();
+        }
         return filmStorage.getRecommendations(userId);
+    }
+
+    public boolean isLegalUserId(Long userId) {
+        return getById(userId) != null;
+    }
+
+    private boolean isIncorrectId(Long id) {
+        return id == null || id <= 0;
+    }
+
+    private boolean isNotValid(User user) {
+        return user.getLogin().contains(" ")
+                || user.getBirthday().isAfter(LocalDate.now());
+    }
+
+    private void changeNameToLogin(User user) {
+        if (user.getName() == null || user.getName().isEmpty() || user.getName().isBlank()) {
+            log.info("Changed blank user name to user login {}", user.getLogin());
+            user.setName(user.getLogin());
+        }
     }
 }
